requirements.txt
addict==2.4.0
appdirs==1.4.3
certifi==2021.5.30
charset-normalizer==2.0.1
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.1
joblib==1.0.1
kaggle==1.5.12
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mmcls @ file:///tmp/tmpo_hanv41/mmclassification
mmcv-full==1.3.9
model-index==0.1.11
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.3.56
openmim @ git+https://github.com/open-mmlab/mim.git@1bb17cd47d5573f963542395d0f18c82b6d67710
ordered-set==4.0.2
pandas==1.3.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==2.2.0
tifffile==2021.7.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.0.10
yapf==0.31.0
zipp==3.5.0
